+++ libsolidity/formal/SMTEncoder.cpp
@@ -888,6 +888,15 @@ bool SMTEncoder::visit(MemberAccess const& _memberAccess)
 				IntegerType const& integerType = dynamic_cast<IntegerType const&>(*magicType->typeArgument());
 				defineExpr(_memberAccess, memberName == "min" ? integerType.minValue() : integerType.maxValue());
 			}
+			else if (memberName == "interfaceId")
+			{
+				// TODO: move this calculation into ContractDefinition and share with ExpressionCompiler
+				ContractDefinition const& contract = dynamic_cast<ContractType const&>(*magicType->typeArgument()).contractDefinition();
+				uint64_t result{0};
+				for (auto const& function: contract.interfaceFunctionList(false))
+					result ^= fromBigEndian<uint64_t>(function.first.ref());
+				defineExpr(_memberAccess, result);
+			}
 			else
 				// NOTE: supporting name, creationCode, runtimeCode would be easy enough, but the bytes/string they return are not
 				//       at all useable in the SMT checker currently